1-21 完成视频上传功能 1. 【完善】common包：VideoMergeUtils.mergeChunks，完善了这个方法。这个方法可以合并多个文件，采用了零拷贝技术，性能很高。 2. 【新增】common包：FFmpegUtil.getVideoDuration，用于获取视频时长。采用ffmpeg。 3. 【新增】common包：FFmpegUtil.getVideoEncoding，用于获取视频编码格式。采用ffmpeg。 4. 【新增】common包：FFmpegUtil.convert2Mp4，用于将任何编码类型视频转化为H.264编码的mp4视频 5. 【新增】common包：FFmpegUtil.convertVideo2Ts，用于将视频转换为分片的TS格式 6. 【补交】补交1个pom.xml（nilo-mq-consumer包的），上次提交忘记交了 7. 【完善】nilomqconsumer包：VideoTransCodingConsumer.receiveMessage，完善了视频文件处理的全流程，相关流程整理在了这个方法的注释上 8. 【新增】nilomqconsumer包：VideoTransCodingConsumer.convertVideo2Ts，配合上面那个方法的一个私有方法，用于在【将视频转化为TS分片文件】的过程中对文件格式进行校验，并与新增的FFmpegUtil方法进行整合 9. 【修正】将一些common下的QO的类型进行改正，例如com.neon.nilocommon.entity.query.VideoInfoFileUploadQuery.transferResult代表文件转码结果，对应mysql数据表的tinybit，这里应该对应Short，因此从Integer修改为Short。由于这个部分是自动生成的，所以此处为人工纠错。 10. 【新增】common包：Constants，这个类下增加了一些常量定义

+++ nilo-common/src/main/java/com/neon/nilocommon/entity/query/VideoInfoFileUploadQuery.java
@@ -59,12 +59,12 @@ public class VideoInfoFileUploadQuery extends BaseQuery
     /**
      * 0:无更新 1:有更新
      */
-    private Integer updateType;
+    private Short updateType;
 
     /**
      * 0:转码中 1:转码成功 2:转码失败
      */
-    private Integer transferResult;
+    private Short transferResult;
 
     /**
      * 持续时间（秒）
